Switch to FrameTime-based update

+++ Celarix.Starfall/Layout/Atria/FrameTime.cs
@@ -1,6 +1,6 @@
 namespace Celarix.Starfall.Layout.Atria;
 
 /// <summary>
-/// A single frame of Atria execution. This is introduced now for future update-loop migration.
+/// A single frame of Atria execution.
 /// </summary>
-public readonly record struct FrameTime(long Number, TimeSpan Elapsed, TimeSpan Delta);
+public readonly record struct FrameTime(int Number, TimeSpan Elapsed, TimeSpan Delta);
